backend/requirements.txt
fastapi>=0.122.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
alembic>=1.13.1,<2.0.0
psycopg[binary]>=3.1.19,<4.0.0
pydantic-settings>=2.3.3,<3.0.0
taskiq>=0.11.18,<1.0.0
taskiq-redis>=1.0.0,<2.0.0
redis>=5.2.0,<6.0.0
httpx>=0.27.0,<1.0.0
